Server and Client connection
the server is connected at local host and port 5555 for testing you need to use your own ip in "player1.py", "player4.py" and "Client.cs" in the variable called "ip"

+++ 8-bit Dungeon - INternet Explorers/Assets/Scripts/CreateCharacter.cs
@@ -35,8 +35,8 @@ public class CreateCharacter : MonoBehaviour
             character.title = "Frog";
         }
         character.init();
-        currentHp.value = character.stats["MaxHP"].value;
-        currentRecourse.value = character.stats["MaxRecourse"].value;
+        currentHp.value = (int)character.stats["MaxHP"].value;
+        currentRecourse.value = (int)character.stats["MaxRecourse"].value;
         SceneManager.LoadScene("CitySample");
     }
 }
